+++ examples/store/ref_reducer.rs
@@ -1,13 +1,13 @@
 use flurx::reducer::RefReducer;
+use flurx::Scheduler;
 use flurx::selector::wait;
 use flurx::store::Store;
 
 #[tokio::main]
 async fn main() {
     let mut store = Store::<String>::default();
-    let mut reducer = RefReducer::new(&mut store);
-
-    reducer.schedule(|task| async move {
+    let mut scheduler  = Scheduler::new();
+    scheduler.schedule(|task| async move {
         println!("*** Start ***");
         task.will(wait::until(|state: &String| {
             println!("state: {state}");
@@ -16,6 +16,7 @@ async fn main() {
             .await;
         println!("*** Finish ***");
     });
+    let mut reducer = RefReducer::new(&mut store, scheduler);
 
     for i in 0..=20 {
         reducer.dispatch(|mut state: String| {