+++ src/reducer/own.rs
@@ -4,6 +4,7 @@ use core::mem;
 use crate::dispatch::Dispatch;
 use crate::prelude::ReactiveTask;
 use crate::reducer::base::ReducerInner;
+use crate::Scheduler;
 use crate::store::Store;
 
 #[repr(transparent)]
@@ -15,8 +16,8 @@ impl<'state, 'future, State> Reducer<'state, 'future, State>
         State: Clone + 'state + 'future + Default
 {
     #[inline]
-    pub fn new(store: &'state mut Store<State>) -> Reducer<'state, 'future, State> {
-        Self(ReducerInner::new(store))
+    pub fn new(store: &'state mut Store<State>, scheduler: Scheduler<'state, 'future, State>) -> Reducer<'state, 'future, State> {
+        Self(ReducerInner::new(store, scheduler))
     }
 }
 
@@ -43,5 +44,15 @@ impl<'state, 'future, State> Reducer<'state, 'future, State>
         let new_state = dispatch.dispatch(state);
         *self.0.store.ref_mut() = new_state.clone();
         self.0.scheduler.run(new_state).await;
+    }
+
+    #[inline]
+    #[cfg(feature = "sync")]
+    pub fn dispatch_sync<D>(&mut self, dispatch: D)
+        where
+            D: Dispatch<State>
+    {
+        use async_compat::CompatExt;
+        pollster::block_on(self.dispatch(dispatch).compat());
     }
 }