+++ src/reducer/ref.rs
@@ -4,6 +4,7 @@ use core::mem;
 use crate::dispatch::Dispatch;
 use crate::prelude::ReactiveTask;
 use crate::reducer::base::ReducerInner;
+use crate::Scheduler;
 use crate::store::Store;
 
 #[repr(transparent)]
@@ -15,8 +16,8 @@ impl<'state, 'future, State> RefReducer<'state, 'future, State>
         State: 'state + 'future
 {
     #[inline]
-    pub fn new(store: &'state mut Store<State>) -> RefReducer<'state, 'future, State> {
-        Self(ReducerInner::new(store))
+    pub fn new(store: &'state mut Store<State>, scheduler: Scheduler<'state, 'future, &State>) -> RefReducer<'state, 'future, State> {
+        Self(ReducerInner::new(store, scheduler))
     }
 }
 
@@ -42,5 +43,16 @@ impl<'state, 'future, State> RefReducer<'state, 'future, State>
 
         *self.0.store.ref_mut() = dispatch.dispatch(state);
         self.0.scheduler.run(self.0.store.read_ref()).await;
+    }
+    
+    
+    #[inline]
+    #[cfg(feature = "sync")]
+    pub fn dispatch_sync<D>(&mut self, dispatch: D)
+        where
+            D: Dispatch<State>
+    {
+        use async_compat::CompatExt;
+        pollster::block_on(self.dispatch(dispatch).compat());
     }
 }