+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.3
+
+### Futures
+
+- Add `Reducer::run_sync`. To use it, enable the `sync` feature.
+
+### Update
+
+- Change type [`ReactiveTask::run`] to [`impl Future<Output=Out> + 'state`]
+
 ## 0.1.2
 
 ### Futures